requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
langgraph
langchain
langchain-openai
langchain-core
mcp
marker-pdf>=1.10.2
surya-ocr>=0.17.1
torchvision
torchaudio
python-magic
rapidfuzz
simple-salesforce
requests
chromadb
sentence-transformers
python-docx
aiohttp
pymongo
